Use union rules to make plugin extensible & split out rules
Changes the pants-docker plugin to use UnionRules to allow other
targets/functionality to be added to the dockerization code by
allowing users to define a FieldSet to extract from targets & define a
rule that transforms that FieldSet into a DockerComponent object which
consists of files to be added to the docker context, and a list of
lines which will be run in the DockerFile

Regorganizes the docker plugin code so that different components are
split into their own files for ease of reading and to separate aspects
that do not need to know of each other. Also renames some files to
be more descriptive and hopefully match the rest of the pants
codebase better (i.e. dockerize.py -> package.py, fields.py ->
target.py)