update to pants 2.13
This PR updates the pants plugin to be compatible with pants version
2.13

This involved some minor changes, and a slightly more major change in
response to how lockfiles & resolves are handled.

This plugin won't currently support the 'resolves' feature, which uses
the pex lockfile format.

If we figure out how to install pip dependencies using the constraints
specified in the pex lockfile we will be able to use support repos
that have 'enable_resolves = true'

Right now enable_resolves = True will fail with an execution error.

Removes reliance on the no-longer extant PantsRequirementsFile, which
was replaced by a macro, we will now create and activate a virtualenv
if there are any 3rd party python dependencies inferred by the pants
plugin.

For the plugin repository it adds test files for the
python_requirements.py file, & enables linting in ci, and enables
code formatting backends